Convert to running on original HTML
There's definitely room for improvement here - the original Javascript
could definitely be optimized, and there's some
original-but-commented-out code that I've included to get it into
the commit history (so that I can recover it if necessary).

And, of course, nothing is cached so it's slow as as slow thing.

And tests are still super-sparse.

Still - functioning! Hooray!